when reading a distibution, freq can very big!
git-svn-id: https://ilk.uvt.nl/svn/trunk/sources/Timbl6@14648 12f355fe-0486-481a-ad91-c297ab22b4e3

+++ src/Instance.cxx
@@ -1133,7 +1133,7 @@ namespace Timbl {
     ValueDistribution *result = 0;
     TargetValue *target;
     string buf;
-    int freq;
+    size_t freq;
     double sw;
     char nextCh;
     is >> nextCh;   // skip {
